--- apps/backend/user/src/verifyMFA.ts
@@ -1,18 +1,0 @@
-import { User } from "./models/user";
-import { TOTP } from "otpauth";
-
-const verifyMFA = async (user: User, totp_secret: string) => {
-  // Create TOTP from stored secret
-  const totp = new TOTP({
-    label: "Martletplace",
-    algorithm: "SHA1",
-    digits: 6,
-    secret: user.totp_secret!,
-  });
-
-  const isValid = totp.validate({ token: totp_secret });
-
-  return isValid === 0;
-};
-
-export { verifyMFA };
